score, board id and user id stores in game_board tabel

+++ db/migrate/20141030125116_add_score_to_game_boards.rb
@@ -1,5 +1,5 @@
 class AddScoreToGameBoards < ActiveRecord::Migration
   def change
-    add_column :game_boards, :score, :integer
+    add_column :game_boards, :score, :decimal
   end
 end
